--- a/Sofa/Component/Constraint/Projective/src/sofa/component/constraint/projective/FixedPlaneConstraint.inl
+++ b/Sofa/Component/Constraint/Projective/src/sofa/component/constraint/projective/FixedPlaneConstraint.inl
@@ -308,13 +308,10 @@ void FixedPlaneConstraint<DataTypes>::draw(const VisualParams* vparams)
 }
 
 template<class DataTypes>
-bool FixedPlaneConstraint<DataTypes>::isPointInPlane(Coord p) const
+bool FixedPlaneConstraint<DataTypes>::isPointInPlane(const Coord& p) const
 {
     const Real d = getVec(p) * getVec(d_direction.getValue());
-    if ((d>d_dmin.getValue()) && (d<d_dmax.getValue()))
-        return true;
-    else
-        return false;
+    return d > d_dmin.getValue() && d < d_dmax.getValue();
 }
 
 } // namespace sofa::component::constraint::projective