--- a/Sofa/Component/Constraint/Projective/src/sofa/component/constraint/projective/FixedTranslationConstraint.h
+++ b/Sofa/Component/Constraint/Projective/src/sofa/component/constraint/projective/FixedTranslationConstraint.h
@@ -92,7 +92,7 @@ public:
 protected:
     template <class DataDeriv>
     void projectResponseT(DataDeriv& dx,
-        std::function<void(DataDeriv&, const unsigned int)> clear);
+        const std::function<void(DataDeriv&, const unsigned int)>& clear);
 
 };