--- a/Sofa/Component/Constraint/Projective/src/sofa/component/constraint/projective/HermiteSplineConstraint.h
+++ b/Sofa/Component/Constraint/Projective/src/sofa/component/constraint/projective/HermiteSplineConstraint.h
@@ -116,8 +116,8 @@ public:
 
 protected:
     template <class DataDeriv>
-    void projectResponseT(const core::MechanicalParams* mparams /* PARAMS FIRST */, DataDeriv& dx,
-        std::function<void(DataDeriv&, const unsigned int)> clear = [](auto& dx, const unsigned int index) {dx[index].clear(); });
+    void projectResponseT(DataDeriv& dx,
+        std::function<void(DataDeriv&, const unsigned int)> clear);
 
 };