… Sofa/Component/Constraint/Projective/src/sofa/component/constraint/projective/HermiteSplineConstraint.inl
… Sofa/Component/Constraint/Projective/src/sofa/component/constraint/projective/LinearMovementConstraint.h
… Sofa/Component/Constraint/Projective/src/sofa/component/constraint/projective/LinearMovementConstraint.inl
… Sofa/Component/Constraint/Projective/src/sofa/component/constraint/projective/OscillatorConstraint.h
… Sofa/Component/Constraint/Projective/src/sofa/component/constraint/projective/OscillatorConstraint.inl
… Sofa/Component/Constraint/Projective/src/sofa/component/constraint/projective/ParabolicConstraint.h
… Sofa/Component/Constraint/Projective/src/sofa/component/constraint/projective/ParabolicConstraint.inl
… Sofa/Component/Constraint/Projective/src/sofa/component/constraint/projective/PartialFixedConstraint.h
… Sofa/Component/Constraint/Projective/src/sofa/component/constraint/projective/PartialFixedConstraint.inl
… Sofa/Component/Constraint/Projective/src/sofa/component/constraint/projective/PartialLinearMovementConstraint.h
… Sofa/Component/Constraint/Projective/src/sofa/component/constraint/projective/PartialLinearMovementConstraint.inl
… Sofa/Component/Constraint/Projective/src/sofa/component/constraint/projective/PatchTestMovementConstraint.h
… Sofa/Component/Constraint/Projective/src/sofa/component/constraint/projective/PatchTestMovementConstraint.inl
… Sofa/Component/Constraint/Projective/src/sofa/component/constraint/projective/ProjectToPointConstraint.inl
… Sofa/Component/Constraint/Projective/src/sofa/component/constraint/projective/SkeletalMotionConstraint.h
… Sofa/Component/Constraint/Projective/src/sofa/component/constraint/projective/SkeletalMotionConstraint.inl
… Sofa/Component/Haptics/src/sofa/component/haptics/LCPForceFeedback.inl
… Sofa/Component/Haptics/tests/scenes/ToolvsFloorCollision_test.scn
… Sofa/Component/Mapping/NonLinear/src/sofa/component/mapping/nonlinear/DistanceFromTargetMapping.inl
… Sofa/Component/Mapping/NonLinear/src/sofa/component/mapping/nonlinear/DistanceMapping.inl
… Sofa/Component/Mapping/NonLinear/src/sofa/component/mapping/nonlinear/SquareDistanceMapping.inl
… Sofa/Component/Mapping/NonLinear/src/sofa/component/mapping/nonlinear/SquareMapping.inl
… Sofa/framework/DefaultType/CMakeLists.txt
… Sofa/framework/DefaultType/src/sofa/defaulttype/CompressedRowSparseMatrixConstraint.cpp
… Sofa/framework/DefaultType/src/sofa/defaulttype/CompressedRowSparseMatrixConstraint.h
… Sofa/framework/DefaultType/src/sofa/defaulttype/CompressedRowSparseMatrixConstraintEigenUtils.h
… Sofa/framework/DefaultType/src/sofa/defaulttype/MapMapSparseMatrixEigenUtils.h
… Sofa/framework/DefaultType/src/sofa/defaulttype/RigidDeriv.h
… Sofa/framework/DefaultType/src/sofa/defaulttype/RigidTypes.h
… Sofa/framework/DefaultType/src/sofa/defaulttype/VecTypes.h