and replace MapMapSparseMatrix

Continuation of #3515 

Breaking because one or two functions have been deleted (for perf reasons), especially the batch clear function.

Some benchs, where constraints are supposedly the bottleneck

caduceus.scn
```
before: 5000 iterations done in 48.2532 s ( 103.62 FPS).
after:  5000 iterations done in 22.042 s ( 226.839 FPS).
```

TorusFall.scn
```
before: 5000 iterations done in 47.5045 s ( 105.253 FPS).
after:  5000 iterations done in 37.8749 s ( 132.014 FPS).
```

3instruments_collis.scn
```
before: 5000 iterations done in 63.3632 s ( 78.9101 FPS).
after:  5000 iterations done in 53.7899 s ( 92.9542 FPS).
```


______________________________________________________

By submitting this pull request, I acknowledge that  
**I have read, understand, and agree [SOFA Developer Certificate of Origin (DCO)](https://github.com/sofa-framework/sofa/blob/master/CONTRIBUTING.md#sofa-developer-certificate-of-origin-dco)**.
______________________________________________________

**Reviewers will merge this pull-request only if**  
- it builds with SUCCESS for all platforms on the CI.
- it does not generate new warnings.
- it does not generate new unit test failures.
- it does not generate new scene test failures.
- it does not break API compatibility.
- it is more than 1 week old (or has fast-merge label).