@alxbilger 2023-06-02
How come the caduceus runs in only 100 fps?