@fredroy 2023-06-14
> I don't understand why we have so much complexity around `projectResponseT`. It seems we don't need the template since we call it only once. Same for the lambda.

Not everywhere, sometimes the templated projectResponseT is called with VecDeriv (aka Vec<Deriv> and MatrixDeriv (aka CRS))
But it could simplified where this statement does not apply though.